requirements.txt
mlflow==2.12.1
scikit-learn==1.3.2
python-dotenv==1.1.1
boto3==1.37.18